+++ app/Http/Controllers/Degree/DegreeController.php
@@ -31,6 +31,7 @@ class DegreeController extends Controller
     }
     public function create()
     {
+        abort_if(!$this->setSlug('addDegree')->hasPermission(), 403, 'You don\'t have permission!');
         return \view('backend.pages.degree.create');
     }
     public function validate_inputs(Request $request)
@@ -39,6 +40,7 @@ class DegreeController extends Controller
     }
     public function store(DegreeAddRequest $request)
     {
+        abort_if(!$this->setSlug('addDegree')->hasPermission(), 403, 'You don\'t have permission!');
         try{
             $response = $this->degreeService->createDegree($request->validated());
             if (is_int($response)) {
@@ -52,6 +54,7 @@ class DegreeController extends Controller
     }
     public function edit($id )
     {
+        abort_if(!$this->setSlug('editDegree')->hasPermission(), 403, 'You don\'t have permission!');
         $degree_info = $this->degreeService->getDegree($id);
         if($degree_info=="Restore first")
             return redirect()->back()->with('error', $degree_info);
@@ -63,6 +66,7 @@ class DegreeController extends Controller
     }
     public function update(DegreeEditRequest $request)
     {
+        abort_if(!$this->setSlug('editDegree')->hasPermission(), 403, 'You don\'t have permission!');
         try{
             if($this->degreeService->edit($request->validated()))
                 return redirect('degree/')->with('success', "Degree updated successfully.");
@@ -73,6 +77,7 @@ class DegreeController extends Controller
     }
     public function delete($id)
     {
+        abort_if(!$this->setSlug('manageDegree')->hasPermission(), 403, 'You don\'t have permission!');
         try{
             if($this->degreeService->delete($id))
                 return redirect('degree/')->with('success', "Degree deleted successfully.");
@@ -83,6 +88,7 @@ class DegreeController extends Controller
     }
     public function restore($id)
     {
+        abort_if(!$this->setSlug('manageDegree')->hasPermission(), 403, 'You don\'t have permission!');
         try{
             if($this->degreeService->restore($id))
                 return redirect('degree/')->with('success', "Degree restored successfully.");
